// admin-next/src/domains/inventory/products/modals/EditSeoModal/types.ts
import type { IMediaFile } from "../../mocks/types";

export interface ISeoFormValues {
  seoTitle: string;
  seoDescription: string;
  ogTitle: string;
  ogDescription: string;
  ogImage: IMediaFile | null;
}

export interface IPreviewProps {
  values: ISeoFormValues;
  productTitle: string;
  baseUrl: string;
  slug?: string;
}